requirements.txt
PyPDF2==3.0.1
pymupdf==1.24.5
Pillow==10.3.0
pandas==2.2.2
pyinstaller==6.8.0
